requirements.txt
appnope==0.1.0
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.0.11
Django==1.11.4
django-bootstrap-pagination==1.6.2
django-ckeditor==5.3.0
django-embed-video==1.1.2
django-js-asset==0.1.1
djangorestframework==3.6.3
entrypoints==0.2.2
html5lib==0.9999999
idna==2.5
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
mongoengine==0.13.0
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.3.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==3.4.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.18.3
selenium==3.4.2
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
urllib3==1.22
wcwidth==0.1.7
widgetsnbextension==1.2.6
